# QuadTree Performance Tuning