--- a/src/DtronixCommon.Tests/Collections/Trees/QuadTreeTests.tt
+++ b/src/DtronixCommon.Tests/Collections/Trees/QuadTreeTests.tt
@@ -53,8 +53,8 @@ public class <#=config.ClassName#>Tests : QuadTreeTestBase
         qt.Insert(0, 0, 0, 0, item);
         qt.Insert(0, 0, 0, 0, item2);
 
-        Assert.AreEqual(0, item.QuadTreeId);
-        Assert.AreEqual(1, item2.QuadTreeId);
+        Assert.That(item.QuadTreeId, Is.EqualTo(0));
+        Assert.That(item2.QuadTreeId, Is.EqualTo(1));
     }
 
     [Test]
@@ -64,7 +64,7 @@ public class <#=config.ClassName#>Tests : QuadTreeTestBase
         var item = new TestQuadTreeItem();
         qt.Insert(0, 0, 0, 0, item);
 
-        Assert.AreEqual(0, item.QuadTreeId);
+        Assert.That(item.QuadTreeId, Is.EqualTo(0));
     }
 
     [Test]
@@ -74,7 +74,7 @@ public class <#=config.ClassName#>Tests : QuadTreeTestBase
         var item = new TestQuadTreeItem();
         qt.Insert(50, 50, 60, 60, item);
 
-        Assert.AreEqual(0, item.QuadTreeId);
+        Assert.That(item.QuadTreeId, Is.EqualTo(0));
     }
 
     [Test]
@@ -86,7 +86,7 @@ public class <#=config.ClassName#>Tests : QuadTreeTestBase
 
         var items = qt.Query(0, 0, 5, 5);
 
-        Assert.AreEqual(0, items.Count);
+        Assert.That(items.Count, Is.EqualTo(0));
     }
 
     [Test]
@@ -105,10 +105,8 @@ public class <#=config.ClassName#>Tests : QuadTreeTestBase
         var items = qt.Query(-1000, -1000, 1000, 1000);
         var items2 = qt.Query(-1000, -1000, 1000, 1000);
 
-        Assert.AreEqual(100, items.Count);
-        Assert.AreEqual(100, items2.Count);
-
-
+        Assert.That(items.Count, Is.EqualTo(100));
+        Assert.That(items2.Count, Is.EqualTo(100));
     }
 }