--- a/src/DtronixCommon.Tests/Threading/DelayedActionTests.cs
+++ b/src/DtronixCommon.Tests/Threading/DelayedActionTests.cs
@@ -46,7 +46,7 @@ public class DelayedActionTests
         });
 
         await delayedAction.InvokeAsync();
-        Assert.IsTrue(await WaitForCompletion(100));
+        Assert.That(await WaitForCompletion(100), Is.True);
     }
 
 
@@ -59,11 +59,11 @@ public class DelayedActionTests
         });
 
         await delayedAction.InvokeAsync();
-        Assert.IsTrue(await WaitForCompletion(100));
+        Assert.That(await WaitForCompletion(100), Is.True);
 
         _tcs = new TaskCompletionSource();
         await delayedAction.InvokeAsync();
-        Assert.IsTrue(await WaitForCompletion(100));
+        Assert.That(await WaitForCompletion(100), Is.True);
 
     }
 
@@ -87,8 +87,8 @@ public class DelayedActionTests
         });
 
 
-        Assert.IsTrue(await WaitForCompletion(1000));
-        Assert.GreaterOrEqual(sw.ElapsedMilliseconds, 80);
+        Assert.That(await WaitForCompletion(1000), Is.True);
+        Assert.That(sw.ElapsedMilliseconds, Is.GreaterThanOrEqualTo(80));
     }
 
     [Test]
@@ -106,6 +106,6 @@ public class DelayedActionTests
             await delayedAction.InvokeAsync(new ArgsValue(1));
         });
 
-        Assert.IsTrue(await WaitForCompletion(150));
+        Assert.That(await WaitForCompletion(150), Is.True);
     }
 }