--- a/src/DtronixCommonBenchmarks/Collections/Trees/QuadTreeBenchmarks.cs
+++ b/src/DtronixCommonBenchmarks/Collections/Trees/QuadTreeBenchmarks.cs
@@ -6,7 +6,6 @@ using DtronixCommon.Collections.Trees;
 namespace DtronixCommonBenchmarks.Collections.Trees;
 
 [MemoryDiagnoser]
-[Config(typeof(FastConfig))]
 public class QuadTreeBenchmarks
 {