Incorporated performance enhancements from quad-tree-overlapping-struct

### Master performance
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.23590
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=8.0.100
  [Host]     : .NET Core 6.0.25 (CoreCLR 6.0.2523.51912, CoreFX 6.0.2523.51912), X64 RyuJIT
  DefaultJob : .NET Core 6.0.25 (CoreCLR 6.0.2523.51912, CoreFX 6.0.2523.51912), X64 RyuJIT


|       Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
|  InsertFloat | 49.72 μs | 0.141 μs | 0.125 μs | 1.0376 |     - |     - |    8736 B |
| InsertDouble | 52.24 μs | 0.136 μs | 0.121 μs | 1.0376 |     - |     - |    8736 B |
|    WalkFloat | 66.66 μs | 0.238 μs | 0.211 μs |      - |     - |     - |         - |
|   WalkDouble | 67.79 μs | 0.246 μs | 0.231 μs |      - |     - |     - |         - |

### PR performance
BenchmarkDotNet v0.13.10, Windows 11 (10.0.23590.1000)
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.100
  [Host]     : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2

| Method       | Mean     | Error    | StdDev   | Gen0   | Allocated |
|------------- |---------:|---------:|---------:|-------:|----------:|
| InsertFloat  | 18.04 μs | 0.025 μs | 0.020 μs | 0.8240 |    7008 B |
| InsertDouble | 18.29 μs | 0.040 μs | 0.036 μs | 0.8240 |    7008 B |
| WalkFloat    | 18.29 μs | 0.034 μs | 0.028 μs |      - |         - |
| WalkDouble   | 18.46 μs | 0.043 μs | 0.036 μs |      - |         - |

Walking is ~3.6x faster and insertion is 2.6x faster!
Part of this is due to PGO being enabled as the PGO portion of this can be seen in the warmup portion of the benchmarks.
```
WorkloadPilot    1: 16 op, 394300.00 ns, 24.6438 us/op
WorkloadPilot    2: 32 op, 768500.00 ns, 24.0156 us/op
WorkloadPilot    3: 64 op, 1521200.00 ns, 23.7688 us/op
WorkloadPilot    4: 128 op, 3060000.00 ns, 23.9062 us/op
WorkloadPilot    5: 256 op, 5930000.00 ns, 23.1641 us/op
WorkloadPilot    6: 512 op, 11752300.00 ns, 22.9537 us/op
WorkloadPilot    7: 1024 op, 23819700.00 ns, 23.2614 us/op
WorkloadPilot    8: 2048 op, 46267500.00 ns, 22.5916 us/op
WorkloadPilot    9: 4096 op, 83309500.00 ns, 20.3392 us/op
WorkloadPilot   10: 8192 op, 151236600.00 ns, 18.4615 us/op
WorkloadPilot   11: 16384 op, 302332400.00 ns, 18.4529 us/op
WorkloadPilot   12: 32768 op, 602806900.00 ns, 18.3962 us/op
```